tests: Use squashfs for static data where possible.
Instead of creating a new InitBasicFS for just about every test,
where a test doesn't need to write to the filesystem it is far
faster to use prepared data on an InitSquashFS instead.

This commit changes as many tests as possible to make this so,
making the tests much faster.